refactor(private): colocate private packages under private
