+++ private/bloom/bloom.go
