+++ private/cipherchunker/chunker.go
