+++ private/cipherfile/dagreader.go
