+++ private/merge.go
@@ -6,7 +6,7 @@ import (
 	"fmt"
 
 	base "github.com/qri-io/wnfs-go/base"
-	ratchet "github.com/qri-io/wnfs-go/ratchet"
+	ratchet "github.com/qri-io/wnfs-go/private/ratchet"
 )
 
 func Merge(ctx context.Context, aNode, bNode base.Node) (result base.MergeResult, err error) {
