+++ private/ratchet/store.go
