+++ wnfs.go
@@ -16,8 +16,8 @@ import (
 	golog "github.com/ipfs/go-log"
 	base "github.com/qri-io/wnfs-go/base"
 	private "github.com/qri-io/wnfs-go/private"
+	ratchet "github.com/qri-io/wnfs-go/private/ratchet"
 	public "github.com/qri-io/wnfs-go/public"
-	ratchet "github.com/qri-io/wnfs-go/ratchet"
 )
 
 var log = golog.Logger("wnfs")
